% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formula.R
\name{add_sub_formula}
\alias{add_sub_formula}
\title{Add or substract molecular formulas.}
\usage{
add_sub_formula(e1, e2, op)
}
\arguments{
\item{e1}{left operand.}

\item{e2}{right operand.}

\item{op}{operator.}
}
\value{
The sum of or difference between two molecular formulas.
}
\description{
Depending on the value of op, molecular formulas e1 and e2 are either added
(op = \code{+}) or substracted (op = \code{-}).
}
\keyword{internal}
