% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{add_sub_mol_form}
\alias{add_sub_mol_form}
\title{Add or subtract molecular formulas.}
\usage{
add_sub_mol_form(op, x, y)
}
\arguments{
\item{op}{Operator.}

\item{x}{Left operand.}

\item{y}{Right operand.}
}
\value{
A molecular formula resulting from the binary operation.
}
\description{
Depending on the value of \code{op}, molecular formulas \code{x} and \code{y} are either added
(op = \code{+}) or subtracted (op = \code{-}).
}
\keyword{internal}
