% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amino_acids}
\alias{amino_acids}
\title{Amino acid residues.}
\format{
A data frame with 22 rows and 4 variables:
\describe{
\item{abbreviation}{single-letter abbreviation}
\item{three_letter}{three-letter abbreviation}
\item{name}{name}
\item{formula}{formula for the residue (i.e., for an amino acid
incorporated into a polypeptide chain)}
}
}
\usage{
amino_acids
}
\description{
A dataset containing abbreviations and formulas
for the 22 proteinogenic amino acids.
}
\keyword{datasets}
