% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formula.R
\name{as.character.mol_form}
\alias{as.character.mol_form}
\title{Convert a molecular formula to a character vector.}
\usage{
\method{as.character}{mol_form}(x, prettify = FALSE, ...)
}
\arguments{
\item{x}{Molecular formula to be converted.}

\item{prettify}{If \code{prettify = TRUE}, include control characters for pretty
printing (blue, positive element counts; red: negative element counts;
gray: empty formulas).}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character representation of the molecular formula
}
\description{
The string representation of a molecular formular is the character \verb{Molecular formula <atoms>}, where \code{atoms} lists all elements and their counts in Hill
order (see internal function \code{\link[=hill_order]{hill_order()}}).
}
\examples{
molecular_formula("N-1 H-3") \%>\% as.character()
molecular_formula("N-1 H-3") \%>\% as.character(prettify = TRUE)

molecular_formula("P+1 O+3 H+1") \%>\% as.character()
molecular_formula("P+1 O+3 H+1") \%>\% as.character(prettify = TRUE)
}
