% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{assemble_proteoforms}
\alias{assemble_proteoforms}
\title{Assemble proteoforms and calculate their masses.}
\usage{
assemble_proteoforms(p, m, mass_set = "average")
}
\arguments{
\item{p}{A protein specification as returned by \code{\link[=define_proteins]{define_proteins()}}.}

\item{m}{A modcom specification as returned by \code{\link[=define_ptm_compositions]{define_ptm_compositions()}}.}

\item{mass_set}{Atomic masses that should be used for mass calculation.}
}
\value{
A data frame containing \eqn{p \times m}{p * m} proteoforms, obtained
by combining information on \eqn{p} proteins (i.e., number of rows in \code{p})
with information on \eqn{m} PTM compositions (i.e., number of rows in \code{m}).
This data frame comprises two columns derived from the input data frames
(\code{protein_name} and \code{modcom_name}, respectively) and two new columns
\code{formula} (overall molecular formula) and \code{mass} (calculated proteoform
masses).
}
\description{
This function combines a set of proteins with a set of PTM compositions
("modcoms") and calculates the masses of the resulting proteoforms (i.e., the
different molecular forms in which the protein product of a single gene can
be found due to changes introduced by posttranslational modifications).
}
\examples{
proteins <- define_proteins(
  system.file("extdata", "mab_sequence.fasta", package = "fragquaxi"),
  .disulfides = 16
)

modcoms <- define_ptm_compositions(sample_modcoms)

assemble_proteoforms(proteins, modcoms)
}
\seealso{
\code{\link[=get_mass]{get_mass()}} for predefined mass sets
}
