% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atomic_masses}
\alias{atomic_masses}
\title{Average and monoisotopic atomic masses.}
\format{
A data frame with 118 rows and 6 variables:
\describe{
\item{element}{chemical symbol}
\item{name}{IUPAC name}
\item{Z}{atomic number (number of protons in the nucleus)}
\item{average}{average atomic mass as defined by IUPAC}
\item{A}{mass number (total number of protons and neutrons in the nucleus)
of the most abundant isotope}
\item{monoisotopic}{atomic mass of the most abundant isotope}
}
}
\source{
\itemize{
\item Average masses: Atomic weights of the elements 2013, Tables 2 and 3
(\url{https://doi.org/10.1515/pac-2015-0305})
\item Isotope abundances: Isotopic compositions of the elements 2013, Table 1
(\url{https://doi.org/10.1515/pac-2015-0503})
\item Monoisotopic masses: The AME16 Atomic Mass Evaluation
(\url{https://www-nds.iaea.org/amdc/ame2016/mass16.txt})
}
}
\usage{
atomic_masses
}
\description{
A dataset containing average and monoisotopic masses
for the chemical elements. Masses are missing in case of elements
for which no stable isotope is known.
}
\keyword{datasets}
