% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{charge}
\alias{charge}
\title{Calculate mass-to-charge ratios.}
\usage{
charge(m, z, charge_agent_mass = NULL)
}
\arguments{
\item{m}{Mass.}

\item{z}{Vector of charges.}

\item{charge_agent_mass}{Mass of the charge carrier. If \code{NULL}, use the
average IUPAC mass of the proton.}
}
\value{
A named vector of mass-to-charge ratios; names correspond to the
values of \code{z}, converted to character.
}
\description{
The function is vectorized over \code{z} and thus may calculate several \eqn{m/z}
values for a single mass.
}
\examples{
charge(150000, z = 24)

charge(150000, z = 20:24)

charge(150000, z = 20:24, charge_agent_mass = 3)
}
