% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{compare_mol_form}
\alias{compare_mol_form}
\title{Generate proxy for comparing molecular formulas.}
\usage{
compare_mol_form(x)
}
\arguments{
\item{x}{Vector of molecular formulas.}
}
\value{
A string vector.
}
\description{
Returns a string vector as proxy for comparing and sorting molecular
formulas. Each string corresponds to a molecular formula and is derived by
pasting the chemical elements and their counts. Counts are padded to ten
digits. Carbon and hydrogen are renamed to 'AC' and 'AH', respectively, in
order to preserve Hill ordering.
}
\keyword{internal}
