% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{define_proteins}
\alias{define_proteins}
\title{Create a protein specification.}
\usage{
define_proteins(..., .disulfides = 0L)
}
\arguments{
\item{...}{FASTA files containing protein sequences. Argument names will be
used as protein names.}

\item{.disulfides}{A vector describing the number of disulfide bridges,
recycled to the number of the given FASTA files. For each disulfide bridge,
two hydrogen atoms are subtracted from the molecular formula.}
}
\value{
A data frame that describes one protein per row and comprises three
columns: \describe{\item{\code{protein_name}}{Name of each protein, derived from
the argument name in \code{...} (if present; otherwise, a consecutive
number).}\item{\code{protein_data}}{Nested column storing the FASTA file name
(\code{file}) and number of disulfides (\code{disulfides}) used for calculating its
formula.}\item{\code{protein_formula}}{Formula calculated from its sequence.}}
}
\description{
This function assembles a protein specification from several FASTA files.
It converts the amino acid sequence stored in each FASTA file to a
molecular formula. If the protein has multiple chains, each chain must be
specified with its own header line, \emph{i.e.}, a line starting with \code{>}. For
each chain, a water molecule is added to the formula.
}
\examples{
mab_sequence <- system.file("extdata", "mab_sequence.fasta",
                            package = "fragquaxi")
define_proteins(mab = mab_sequence)

define_proteins(mab_sequence, .disulfides = 16)
}
