% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{define_proteoforms}
\alias{define_proteoforms}
\title{Create a proteoform specification.}
\usage{
define_proteoforms(ptm_compositions, other_ptms = NULL)
}
\arguments{
\item{ptm_compositions}{A data frame with PTM compositions (see details).}

\item{other_ptms}{A named vector (of strings or molecular formulas) defining
additional PTMs like \verb{c(ptm_1 = correction formula 1, ..., ptm_n = correction formula n)} .}
}
\value{

}
\description{
This function defines proteoforms from post-translational modification (PTM)
compositions.
}
\details{
\subsection{PTM compositions}{

PTM compositions are supplied in a data frame where each row corresponds to a
single composition. This data frame must comprise a column labeled \code{pfm_name}
with a short description. The remaining columns must be labeled by valid PTM
names (see below) and contain the number of the respective PTM found in the
respective composition.

Valid PTM names include
\itemize{
\item monosaccharides as provided in the data set \link{monosaccharides},
\item PTMs as provided in the data set \link{ptms},
\item user-defined PTMs supplied by the user via the argument \code{other_ptms}.
}
}
}
\examples{
ptm_compositions <- tibble::tribble(
  ~pfm_name,       ~Hex, ~HexNAc, ~Fuc, ~PHOS, ~foo,
  "G0F/G0",           6,       8,    1,    0,     0,
  "G0F/G0F",          6,       8,    2,    0,     0,
  "G0F/G1F",          7,       8,    2,    0,     0,
  "G1F/G1F",          8,       8,    2,    0,     0,
  "G1F/G2F",          9,       8,    2,    0,     0,
  "G2F/G2F",         10,       8,    2,    0,     0,
  "G2F/G2F+P",       10,       8,    2,    1,     0,
  "G2F/G2F+foo",     10,       8,    2,    0,     1,
)

my_ptms <- c(
  foo = "C42",
  bar = "H42"
)

define_proteoforms(ptm_compositions, my_ptms)
}
\seealso{
\link{monosaccharides} and \link{ptms} for predefined monosaccharides and
PTMs, respectively
}
