% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{define_ptm_compositions}
\alias{define_ptm_compositions}
\title{Create a specification of PTM compositions.}
\usage{
define_ptm_compositions(x, other_ptms = NULL)
}
\arguments{
\item{x}{A data frame with PTM compositions (see details).}

\item{other_ptms}{A named vector (of strings or molecular formulas) defining
additional PTMs like \verb{c(ptm_1 = correction formula 1, ..., ptm_n = correction formula n)} .}
}
\value{
A data frame that describes one PTM composition per row and comprises
three columns: \describe{\item{\code{modcom_name}}{Name of each PTM composition
(same as in the input data frame).}\item{\code{modcom_data}}{Nested column
storing the remaining columns from the input data
frame.}\item{\code{modcom_formula}}{Correction formula associated with the
respective PTM composition.}}
}
\description{
This function assembles a post-translational modification (PTM)
specification from PTM compositions.
}
\details{
\subsection{PTM compositions}{

PTM compositions ("modcoms") are supplied in a data frame where each row
corresponds to a single composition. This data frame must comprise a column
labeled \code{modcom_name} with a short description. The remaining columns must be
labeled by valid PTM names (see below) and contain the number of the
respective PTM found in the respective composition.

Valid PTM names include
\itemize{
\item monosaccharides as provided in the data set \link{monosaccharides},
\item PTMs as provided in the data set \link{ptms},
\item user-defined PTMs supplied by the user via the argument \code{other_ptms}.
}
}
}
\examples{
ptm_compositions <- tibble::tribble(
  ~modcom_name, ~Hex, ~HexNAc, ~Fuc, ~PHOS, ~foo,
  "G0F/G0",        6,       8,    1,    0,     0,
  "G0F/G0F",       6,       8,    2,    0,     0,
  "G0F/G1F",       7,       8,    2,    0,     0,
  "G1F/G1F",       8,       8,    2,    0,     0,
  "G1F/G2F",       9,       8,    2,    0,     0,
  "G2F/G2F",      10,       8,    2,    0,     0,
  "G2F/G2F+P",    10,       8,    2,    1,     0,
  "G2F/G2F+foo",  10,       8,    2,    0,     1,
)

my_ptms <- c(
  foo = "C42",
  bar = "H42"
)

define_ptm_compositions(ptm_compositions, my_ptms)
}
\seealso{
\link{monosaccharides} and \link{ptms} for predefined monosaccharides and
PTMs, respectively
}
