% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{find_rt_limits}
\alias{find_rt_limits}
\title{Find retention time limits for XIC integration.}
\usage{
find_rt_limits(rt, rt_limits)
}
\arguments{
\item{rt}{Retention times from an \code{mzR} object.}

\item{rt_limits}{Value of the correspondent argument of \code{\link[=quantify_ions]{quantify_ions()}}.}
}
\value{
A data frame with one row per retention time window and the
calculated retention time limits and scan indices in columns \code{rt_min},
\code{rt_max}, \code{scan_min}, and \code{scan_max}, respectively.
}
\description{
This internal function is called by \code{\link[=quantify_ions]{quantify_ions()}}. It
\itemize{
\item converts \code{rt_limits} to a data frame (since it may be a numeric vector with
two elements or a list).
\item replaces any infinite value with the minimum or maximum retention time.
\item finds the scan indices enclosed by the retention time limits.
If a retention time window does not include any scans,
scan indices are \code{NA}.
}
}
\keyword{internal}
