% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formula.R
\name{format.mol_form}
\alias{format.mol_form}
\title{Format a molecular formula.}
\usage{
\method{format}{mol_form}(x, prettify = FALSE, ...)
}
\arguments{
\item{x}{Molecular formula to be converted.}

\item{prettify}{If \code{prettify = TRUE}, include control characters for pretty
printing (blue, positive element counts; red: negative element counts;
gray: empty formulas).}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character representation of the molecular formula
}
\description{
Format a molecular formula as a list of all elements and their counts in
\link[=hill_order]{Hill order}.
}
\examples{
molecular_formula("N-1 H-3") \%>\% format()
molecular_formula("N-1 H-3") \%>\% format(prettify = TRUE)

molecular_formula("P+1 O+3 H+1") \%>\% format()
molecular_formula("P+1 O+3 H+1") \%>\% format(prettify = TRUE)

# arguments in ... are forwarded to the default format() method
molecular_formula("H2O") \%>\% format(justify = "right", width = 50)
}
