% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragquaxi-package.R
\docType{package}
\name{fragquaxi-package}
\alias{fragquaxi}
\alias{fragquaxi-package}
\title{fragquaxi: Glycoform Quantification via XIC Integration}
\description{
Tools for relative quantification of proteoforms in mass
    spectometric data by integration of extracted ion currents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdl-biosimilars/fragquaxi}
  \item Report bugs at \url{https://github.com/cdl-biosimilars/fragquaxi/issues}
}

}
\author{
\strong{Maintainer}: Wolfgang Esser-Skala \email{Wolfgang.Esser-Skala@sbg.ac.at} (\href{https://orcid.org/0000-0002-7350-4045}{ORCID})

Other contributors:
\itemize{
  \item Katharina Müller [artist]
  \item University of Salzburg [copyright holder, funder]
}

}
\keyword{internal}
