% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{fragquaxi-vctrs}
\alias{fragquaxi-vctrs}
\alias{vec_ptype2.mol_form}
\alias{vec_cast.mol_form}
\alias{vec_arith.mol_form}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{mol_form}(x, y, ...)

\method{vec_cast}{mol_form}(x, to, ...)

\method{vec_arith}{mol_form}(op, x, y, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
