% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{hill_order}
\alias{hill_order}
\title{Sort a character vector of chemical elements according to the Hill system.}
\usage{
hill_order(el)
}
\arguments{
\item{f}{Character vector to be sorted.}
}
\value{
A character vector sorted in Hill order.
}
\description{
\itemize{
\item If the vector contains carbon and hydrogen,
the result is \code{c("C", "H", others)}.
\item If the vector contains carbon but not hydrogen,
the result is \code{c("C", others)}.
\item If the formula contains neither carbon nor hydrogen,
the result is \code{c(others)}. In any case, \code{others} is sorted alphabetically.
}
}
\keyword{internal}
