% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{ionize}
\alias{ionize}
\title{Calculate mass-to-charge ratios for proteoform ions.}
\usage{
ionize(proteoforms, charge_states, ppm = 400, charge_agent_mass = NULL)
}
\arguments{
\item{proteoforms}{A data frame describing proteoform masses (see details).}

\item{charge_states}{Vector of charge states for which m/z values should be
calculated.}

\item{ppm}{Width of the m/z window in parts per million.}

\item{charge_agent_mass}{Mass of the charge carrier. If \code{NULL}, use the
average IUPAC mass of the proton.}
}
\value{
A data frame containing all columns of the supplied data frame plus
four additional columns: \code{z} (charge state), \code{mz} (exact mass-to-charge
ratio), \code{mz_min}, and \code{mz_max}. The latter two columns represent lower and
upper tolerances for the mass-to-charge ratio and are used as integration
boundaries by \code{\link[=quantify_ions]{quantify_ions()}}. If the supplied data frame has \eqn{r}
rows and \eqn{c} charge states were specified, the resulting data frame has
\eqn{r * c} rows.
}
\description{
For each value \eqn{z} in \code{charge_states}, the function calculates
\itemize{
\item the corresponding mass-to-charge ratio \eqn{m / z + mc}
(\eqn{mc}: charge agent mass);
\item upper and lower tolerances for this ratio, so that these values
form a window of width \eqn{mz * ppm / 1000000} centered at \eqn{mz}.
}
}
\details{
The data frame passed to \code{proteoforms} must contain at least one column named
\code{mass}; each of its value represents the mass of a single proteoform. The
data frame may contain any number of additional columns, which will be
returned unchanged.
}
\examples{
masses <- data.frame(name = c("p1", "p2"), mass = c(140000, 145000))
ionize(masses, 36L:40L)
}
