% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{load_protein_sequence}
\alias{load_protein_sequence}
\title{Calculate the molecular formula of a protein.}
\usage{
load_protein_sequence(f, disulfides = 0L)
}
\arguments{
\item{f}{FASTA file containing the protein sequence.}

\item{disulfides}{Number of disulfide bridges. For each disulfide bridge, two
hydrogen atoms are subtracted from the molecular formula.}
}
\value{
The molecular formula of the protein.
}
\description{
This function converts the amino acid sequences stored in a FASTA file to a
molecular formula. If the protein has multiple chains, each chain must be
specified with its own header line, \emph{i.e.}, a line starting with \code{>}. For
each chain, a water molecule is added to the formula.
}
\examples{
mab_sequence <- system.file("extdata", "mab_sequence.fasta", package = "fragquaxi")
load_protein_sequence(mab_sequence)

load_protein_sequence(mab_sequence, disulfides = 16)
}
