% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formula.R
\name{molecular_formula}
\alias{molecular_formula}
\alias{+.mol_form}
\alias{-.mol_form}
\alias{*.mol_form}
\title{Generate a molecular formula.}
\usage{
molecular_formula(s, simplify = TRUE)

\method{+}{mol_form}(e1, e2)

\method{-}{mol_form}(e1, e2)

\method{*}{mol_form}(e1, e2)
}
\arguments{
\item{s}{Character vector containing a molecular formula.}

\item{simplify}{If \code{simplify = TRUE}, return a single molecular formula
object if possible (that is, if \code{s} contains a single value).}

\item{e1}{left operand}

\item{e2}{right operand}
}
\value{
If \code{s} is a scalar and \code{simplify = TRUE}, the function returns a
single molecular formula object. Otherwise, it returns a list of molecular
formulas.
}
\description{
Convert a character vector to a molecular formula. \code{molecular_formula()} is
vectorized over \code{s}, which allows its usage in functions such as
\code{\link[dplyr:mutate]{dplyr::mutate()}}.
}
\details{
A molecular formula object is a named integer vector with class \code{"mol_form"},
whose names (values) correspond to element symbols (counts).
}
\section{Methods (by generic)}{
\itemize{
\item \code{+}: Add two molecular formulas \code{e1} and \code{e2}.

\item \code{-}: Subtract molecular formula \code{e2} from \code{e1}.

\item \code{*}: Multiply a molecular formula \code{e1} with an
integer \code{e2} (or conversely). The multiplier is always converted to an
integer.
}}

\examples{
molecular_formula("C35 H48 Cl1 N3 O10 S1")

# spaces and counts with value 1 are optional
molecular_formula("C35H48ClN3O10S")

# counts may be negative
molecular_formula("H-3 N-1")

# returns a single molecular formula object
molecular_formula("H2SO4")

# returns a list with a single element
molecular_formula("H2SO4", simplify = FALSE)

# returns a list with one element for each parsed string
molecular_formula(c("H2 O", "C6 H12 O6"))
}
