% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monosaccharides}
\alias{monosaccharides}
\title{Common monosaccharide residues.}
\format{A data frame with 83 rows and 5 variables:
\describe{
\item{abbreviation}{abbreviation}
\item{type}{carbohydrate type}
\item{formula}{formula for the residue (i.e., for a monosaccharide
incorporated into a glycan)}
\item{short_name}{short name}
\item{systematic_name}{IUPAC name}
}}
\source{
\itemize{
\item Abbreviations and names: Symbol Nomenclature for Glycans (SNFG),
2019 edition, Table 3 (\url{https://www.ncbi.nlm.nih.gov/glycans/snfg.html})
\item Formulas: PubChem (\url{https://pubchem.ncbi.nlm.nih.gov/})
}
}
\usage{
monosaccharides
}
\description{
A dataset describing molecular compositions for common monosaccharide
residues.
}
\keyword{datasets}
