% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{mul_div_mol_form}
\alias{mul_div_mol_form}
\title{Multiply or divide molecular formulas.}
\usage{
mul_div_mol_form(op, x, y)
}
\arguments{
\item{op}{Operator.}

\item{x}{Left operand.}

\item{y}{Right operand.}
}
\value{
A molecular formula resulting from the binary operation.
}
\description{
Depending on the value of \code{op}, molecular formula \code{x} is either multiplied by
numeric \code{y} (op = \code{*}) or integer divided (op = \code{\%/\%} or \code{/}). \code{y} is always
cast to integer.
}
\keyword{internal}
