% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{new_mass_set}
\alias{new_mass_set}
\title{Create a new mass set.}
\usage{
new_mass_set(masses, inherits_from = NULL)
}
\arguments{
\item{masses}{Named numeric vector of the form \code{c(element_1 = mass_1, ..., element_n = mass_n)}.}

\item{inherits_from}{Mass set from which masses should be inherited if they
are not specified in \code{masses}. Either \code{"average"}, \code{"monoisotopic"} or a
variable containing a mass set previously defined by \code{new_mass_set()}.}
}
\value{
A named numeric vector containing supplied and inherited element
masses.
}
\description{
The custom mass set will be available in subsequent calls of \code{get_mass()}.
You may specify masses of selected elements only. In this case, however,
calculating the mass of a formula that contains an element whose mass was not
specified will return \code{NA}.
}
\examples{
my_masses <- new_mass_set(c(H = 10, C = 100, O = 1))
molecular_formula("C6 H8 O5") \%>\% get_mass(my_masses)

# returns NA since my_masses does not specify the mass of sulfur
molecular_formula("H2 S O4") \%>\% get_mass(my_masses)

# inherit masses from another mass set to provide default values
my_masses_complete <- new_mass_set(
  c(H = 10, C = 100, O = 1),
  inherits_from = "average"
)
molecular_formula("H2 S O4") \%>\% get_mass(my_masses_complete)
}
\seealso{
\code{\link[=get_mass]{get_mass()}} for calculating molecular masses, \code{\link{atomic_masses}} for
predefined mass sets \code{"average"} and \code{"monoisotopic"}
}
