% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ions}
\alias{plot_ions}
\title{Visualize integration boundaries for proteoform ions.}
\usage{
plot_ions(
  ms_data,
  ions,
  scans = NULL,
  rt_limits = NULL,
  xlim = NULL,
  plot_elements = c("mz", "mz_min", "mz_max", "ion_labels"),
  filter_ms1 = TRUE
)
}
\arguments{
\item{ms_data}{Mass spectrometry data stored in an mzR object as returned by
\code{\link[mzR:openMSfile]{mzR::openMSfile()}}.}

\item{ions}{A data frame describing mass-to-charge ratios of proteoform ions
as returned by \code{\link[=ionize]{ionize()}}.}

\item{scans}{Scan numbers to show (vector of integers). If \code{NULL}, show all
spectra in the file.}

\item{rt_limits}{A numeric vector with two elements specifying retention time
limits (in seconds) for selecting scans. If values are specified for both
\code{scans} and \code{rt_limits}, the function issues a warning and ignores the
value of the latter.}

\item{xlim}{x-axis limits, given as \code{c(lower, upper)}. If \code{NULL}, use the
default scale range.}

\item{plot_elements}{Character vector specifying which plot elements should
be displayed: \describe{ \item{\code{"mz"}}{Solid red vertical lines indicating
m/z values.} \item{\code{"mz_min"}, \code{"mz_max"}}{Dashed red vertical lines
indicating lower and upper integration boundaries, respectively.}
\item{\code{"ion_labels"}}{Secondary x-axis, labeled by the names and charges of
the proteoform ions shown.} }}

\item{filter_ms1}{If true, only scans from MS level 1 are selected by the
retention time limits.}
}
\value{
A ggplot object describing the created plot.
}
\description{
This function plots one or several mass spectra overlaid with mass-to-charge
ratios and integration boundaries of proteoform ions.
}
\examples{
ms_data <- mzR::openMSfile(
  system.file("extdata", "mzml", "mab1.mzML", package = "fragquaxi")
)

proteins <- define_proteins(
  system.file("extdata", "mab_sequence.fasta", package = "fragquaxi"),
  .disulfides = 16
)

modcoms <- define_ptm_compositions(sample_modcoms)

pfm_ions <-
  assemble_proteoforms(proteins, modcoms) \%>\%
  ionize(36L)

# show all scans
plot_ions(ms_data, pfm_ions, xlim = c(4100, 4150))

# show selected scans
plot_ions(ms_data, pfm_ions, scans = 126:136, xlim = c(4100, 4150))

# show a single scan
plot_ions(ms_data, pfm_ions, scans = 126, xlim = c(4100, 4150))

# only show selected plot elements
plot_ions(ms_data, pfm_ions,
          scans = 126, xlim = c(4100, 4150),
          plot_elements = c("mz", "ion_labels"))
}
