% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formula.R
\name{print.mol_form}
\alias{print.mol_form}
\title{Print a molecular formula.}
\usage{
\method{print}{mol_form}(x, ...)
}
\arguments{
\item{x}{molecular formula to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function returns the printed molecular formula invisibly.
}
\description{
Printing yields the string \verb{Molecular formula <atoms>}, where \code{atoms} lists
all elements and their counts in \link[=hill_order]{Hill order}.
}
\examples{
molecular_formula("N-1 H-3")

molecular_formula("P+1 O+3 H+1")
}
