% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{print.quaxi}
\alias{print.quaxi}
\title{Printing quantification results.}
\usage{
\method{print}{quaxi}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
The function returns quantification results invisibly.
}
\description{
Print a concise summary of quantification results.
}
