% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptms}
\alias{ptms}
\title{Common post-translational modifications.}
\format{A data frame with 72 rows and 3 variables:
\describe{
\item{abbreviation}{abbreviation}
\item{name}{description}
\item{formula}{correction formula}
}}
\source{
\itemize{
\item Abbreviations and name: ExPASy FindMod tool
(\url{https://web.expasy.org/findmod/findmod_masses.html})
\item Correction formulas: UniProt Controlled vocabulary of
posttranslational modifications (PTM), release 2020_01 of 26-Feb-2020
(\url{https://www.uniprot.org/docs/ptmlist})
}
}
\usage{
ptms
}
\description{
A dataset describing molecular compositions for common post-translational
modifications.
}
\keyword{datasets}
