% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_modcoms}
\alias{sample_modcoms}
\title{Exemplary PTM compositions ("modcoms").}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{modcom_name}{abbreviation}
\item{Hex}{number of hexoses}
\item{HexNAc}{number of N-acetylhexosamines}
\item{Fuc}{number of fucoses}
}
}
\usage{
sample_modcoms
}
\description{
A dataset describing PTM compositions.
}
\keyword{datasets}
