% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{sum_mol_form}
\alias{sum_mol_form}
\title{Sum of molecular formulas.}
\usage{
sum_mol_form(x, ...)
}
\arguments{
\item{x}{Vector of molecular formulas.}

\item{...}{Additional arguments in the call to \code{sum()}.}
}
\value{
The sum of all values present in \code{x}.
}
\description{
Sum of molecular formulas.
}
\keyword{internal}
