% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trapz_limits}
\alias{trapz_limits}
\title{Trapezoidal integration with limits.}
\usage{
trapz_limits(x, y, lower, upper)
}
\arguments{
\item{x}{x-coordinates of points on the x-axis.}

\item{y}{Function values associated with the x-coordinates.}

\item{lower}{Lower integration boundary.}

\item{upper}{Upper integration boundary.}
}
\value{
Area under the curve defined by the (x, y)-tuples.
}
\description{
Compute the area of a function with values y at the points x within given
limits.
}
\examples{
x <- 0:5
y <- c(1, 4, 2, 3, 3, 0)

trapz_limits(x, y, -2, 10)

trapz_limits(x, y, -2, -1)

trapz_limits(x, y, -2, .4)

trapz_limits(x, y, .4, .7)

trapz_limits(x, y, .4, 3.6)

trapz_limits(x, y, .4, 7)

trapz_limits(x, y, 6, 7)

\dontrun{
# throws an error
trapz_limits(x, y, 4, 3)
}
}
