% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ions.R
\name{weigh}
\alias{weigh}
\title{Calculate proteoform masses.}
\usage{
weigh(proteins, proteoforms, mass_set = "average")
}
\arguments{
\item{proteins}{A protein specification as returned by \code{\link[=define_proteins]{define_proteins()}}.}

\item{proteoforms}{A proteoform specification as returned by
\code{\link[=define_proteoforms]{define_proteoforms()}}.}

\item{mass_set}{Atomic masses that should be used for mass calculation.}
}
\value{
A data frame containing \eqn{p \times n}{p * n} proteoforms, obtained
by combining information on \eqn{p} proteins (as specified in \code{proteins})
with information on \eqn{n} proteoforms (as specified in \code{proteoforms}).
This data frame comprises two columns derived from the input data frames
(\code{protein_name} and \code{pfm_name}, respectively) and two new columns \code{formula}
(overall molecular formula) and \code{mass} (calculated proteoform masses).
}
\description{
This function combines a set of proteins with a set of proteoforms and
calculates the masses of the resulting proteoforms (i.e., the different
molecular forms in which the protein product of a single gene can be found
due to changes introduced by posttranslational modifications).
}
\examples{
proteins <- define_proteins(
  system.file("extdata", "mab_sequence.fasta", package = "fragquaxi"),
  .disulfides = 16
)

proteoforms <- define_proteoforms(
  tibble::tribble(
    ~pfm_name,       ~Hex, ~HexNAc, ~Fuc,
    "G0F/G0",           6,       8,    1,
    "G0F/G0F",          6,       8,    2,
    "G0F/G1F",          7,       8,    2,
    "G1F/G1F",          8,       8,    2,
    "G1F/G2F",          9,       8,    2,
    "G2F/G2F",         10,       8,    2,
  )
)

weigh(proteins, proteoforms)
}
\seealso{
\code{\link[=get_mass]{get_mass()}} for predefined mass sets
}
