% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{as_tibble.quaxi}
\alias{as_tibble.quaxi}
\title{Coerce quantification results to a data frame.}
\usage{
\method{as_tibble}{quaxi}(x, abundance_col = "abundance_data", xic_col = "xic_data", ...)
}
\arguments{
\item{x}{A \code{quaxi} object that should be coerced to a data frame.}

\item{abundance_col}{Name of the column that will contain abundance data. If
\code{NULL}, do not include abundance data in the output.}

\item{xic_col}{Name of the column that will contain XIC data. If \code{NULL}, do
not include XIC data in the output.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
The data frame stored in \code{x$ions} without column \code{ion_id} (i.e., a
data frame with one ion per row) and up to two nested columns containing
data on abundances and XICs, respectively.

The former column is labeled according to the value of \code{abundance_col} and
contains the variables \code{rt_min}, \code{rt_max}, \code{scan_min}, \code{scan_max}, and
\code{abundance}.

The latter column is labeled according to the value of \code{xic_col} and
contains the variables \code{scan} (scan index), \code{rt} (corresponding retention
time), and \code{int} (intensity).
}
\description{
This function converts quantification results obtained from \code{\link[=quantify_ions]{quantify_ions()}}
to a nested data frame, which provides data on abundances and extracted ion
currents (XIC) for each quantified ion.
}
\examples{
library(tibble)

ms_data <- mzR::openMSfile(
  system.file("extdata", "mzml", "mab1.mzML", package = "fragquaxi")
)

proteins <- define_proteins(
  system.file("extdata", "mab_sequence.fasta", package = "fragquaxi"),
  .disulfides = 16
)

modcoms <- define_ptm_compositions(sample_modcoms)

pfm_ions <-
  assemble_proteoforms(proteins, modcoms) \%>\%
  ionize(36L:40L)

abundances <- quantify_ions(ms_data, pfm_ions, c(300, 350))

# by default, include data on abundances and XICs in the output
as_tibble(abundances)

# rename the column containing abundance data
# (and omit several columns for readability)
as_tibble(abundances, abundance_col = "abundances") \%>\%
  dplyr::select(-formula, -mz_min, -mz_max)

# omit XIC data in the output
as_tibble(abundances, xic_col = NULL) \%>\%
  dplyr::select(-formula, -mz_min, -mz_max)
}
