% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formula.R
\name{get_mass}
\alias{get_mass}
\title{Calculate the mass of a molecular formula.}
\usage{
get_mass(f, mass_set = "average")
}
\arguments{
\item{f}{Molecular formula whose mass should be calculated.}

\item{mass_set}{Atomic masses that should be used for mass calculation.
Either a string specifying a predefined mass set (\code{"average"}, average
atomic masses as defined by IUPAC; \code{"monoisotopic"}, masses of the most
abundant isotope of each element), or a variable containing a custom mass
set defined by \code{\link[=new_mass_set]{new_mass_set()}}.}
}
\value{
A scalar double representing the molecular mass (in dalton).
}
\description{
Calculate the mass of a molecular formula.
}
\examples{
molecular_formula("C6 H10 O5") \%>\% get_mass()

molecular_formula("C6 H10 O5") \%>\% get_mass("monoisotopic")
}
\seealso{
\code{\link[=new_mass_set]{new_mass_set()}} for defining custom mass sets, \code{\link{atomic_masses}}
for predefined mass sets
}
