% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_formula.R
\name{hill_order}
\alias{hill_order}
\title{Sort a molecular formula according to the Hill system.}
\usage{
hill_order(f)
}
\arguments{
\item{f}{Molecular formula to be sorted.}
}
\value{
A molecular formular in Hill order.
}
\description{
\itemize{
\item If the formula contains carbon and hydrogen,
list the number of carbon atoms,
followed by the number of hydrogen atoms,
followed by the remaining elements in alphabetical order.
\item If the formula contains carbon but not hydrogen,
list the number of carbon atoms,
followed by the remaining elements in alphabetical order.
\item If the formula contains neither carbon nor hydrogen,
list all elements in alphabetical order.
}
}
\keyword{internal}
