% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_form.R
\name{molecular_formula}
\alias{molecular_formula}
\alias{is_molecular_formula}
\title{Vector of molecular formulas.}
\usage{
molecular_formula(s = character())

is_molecular_formula(x)
}
\arguments{
\item{s}{Character vector describing molecular formulas.}

\item{x}{An object to test.}
}
\value{
An S3 vector of class \code{mol_form}.
}
\description{
Generate a vector of molecular formulas from a character vector.
}
\examples{
molecular_formula("C35 H48 Cl1 N3 O10 S1")

# spaces and counts with value 1 are optional
molecular_formula("C35H48ClN3O10S")

# counts may be negative
molecular_formula("H-3 N-1")

# the function is vectorized
molecular_formula(c("H2 O", "C6 H12 O6", "C Cl4"))

# ... and handles empty and missing formulas
molecular_formula(c("H2 O", "", "<empty>", NA))
}
