% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ions}
\alias{plot_ions}
\title{Visualize integration boundaries for proteoform ions.}
\usage{
plot_ions(
  ms_data,
  ions,
  scans = NULL,
  rt_limits = NULL,
  xlim = NULL,
  plot_elements = c("mz", "mz_min", "mz_max", "ion_labels")
)
}
\arguments{
\item{ms_data}{Mass spectrometry data stored in an mzR object as returned by
\code{\link[mzR:openMSfile]{mzR::openMSfile()}}.}

\item{ions}{A data frame describing mass-to-charge ratios of proteoform ions
as returned by \code{\link[=ionize]{ionize()}}.}

\item{scans}{Scan numbers to show (vector of integers). If \code{NULL}, show all
spectra in the file.}

\item{rt_limits}{A numeric vector with two elements specifying retention time
limits (in seconds) for selecting scans. If values are specified for both
\code{scans} and \code{rt_limits}, the function issues a warning and ignores the
value of the latter.}

\item{xlim}{x-axis limits, given as \code{c(lower, upper)}. If \code{NULL}, use the
default scale range.}

\item{plot_elements}{Character vector specifying which plot elements should
be displayed: \describe{ \item{\code{"mz"}}{Solid red vertical lines indicating
m/z values.} \item{\code{"mz_min"}, \code{"mz_max"}}{Dashed red vertical lines
indicating lower and upper integration boundaries, respectively.}
\item{\code{"ion_labels"}}{Secondary x-axis, labeled by the names and charges of
the proteoform ions shown.} }}
}
\value{
A ggplot object describing the created plot.
}
\description{
This function plots one or several mass spectra overlaid with mass-to-charge
ratios and integration boundaries of proteoform ions.
}
\examples{
ms_data <- mzR::openMSfile(
  system.file("extdata", "mzml", "mab1.mzML", package = "fragquaxi")
)
pfm_ions <- tibble::tribble(
  ~name,       ~Hex, ~HexNAc, ~Fuc,
  "G0F/G0",       6,       8,    1,
  "G0F/G0F",      6,       8,    2,
  "G0F/G1F",      7,       8,    2,
  "G1F/G1F",      8,       8,    2,
  "G1F/G2F",      9,       8,    2,
  "G2F/G2F",     10,       8,    2,
) \%>\%
  calculate_proteoform_masses("C6464 H9950 N1706 O2014 S44") \%>\%
  ionize(36L)

# show all scans
plot_ions(ms_data, pfm_ions, xlim = c(4100, 4150))

# show selected scans
plot_ions(ms_data, pfm_ions, scans = 126:136, xlim = c(4100, 4150))

# show a single scan
plot_ions(ms_data, pfm_ions, scans = 126, xlim = c(4100, 4150))

# only show selected plot elements
plot_ions(ms_data, pfm_ions,
          scans = 126, xlim = c(4100, 4150),
          plot_elements = c("mz", "ion_labels"))
}
